+++ CHANGELOG.md
@@ -1,6 +1,5 @@
 # 0.1.6
 * [ ] Separate **core library** out from this extension.
-* [ ] Support Task List (TODO) for pandoc parser. 
-* [ ] Add `data-line` to task-item.  
+* [ ] Support Task List (TODO) for pandoc parser. Added click event for checkbox.    
 * [ ] Add `Welcome page`.  
 * [ ] Add preview theme `Medium`.